mz305: fix outdated numeric references after skip & squash
After skip & squash the index in the list of stages can change, but references to the stages are not updated.
We fix this by storing the original index on the element and only refering to the stage using it's original index.
For that end the skipUnusedStages was moved to work on the "intermediate" kanikoStages.