feat(login): introduce executor login subcommand for registry login (#407)
* feat(login): introduce executor login subcommand for registry login

This revision introduces a new subcommand to the executor, login, which
offers a familiar interface for logging into a remote registry. This is
an analogue to 'docker login', 'podman login' or 'crane auth login' and
makes it easier to manipulate the .docker/config.json without the
tedious chore of manipulating the file directly in a CI environment.

* cobra cli change global PersistentFlags() to local Flags()

* update copyright

---------

Co-authored-by: Martin Zihlmann <martizih@outlook.com>